changing top peers to TS
