+++ stock-tracker/src/features/topPeers/redux/actionTypes.ts
