--- stock-tracker/src/features/topPeers/redux/actions.js
@@ -1,14 +1,0 @@
-import {
-  ADD_TOP_PEERS,
-  SET_LOADING_PEERS,
-  SET_ERROR_PEERS
-} from "./actionTypes";
-
-export const addTopPeersAction = topPeers => ({
-  type: ADD_TOP_PEERS,
-  payload: topPeers
-});
-
-export const setLoadingPeersAction = () => ({ type: SET_LOADING_PEERS });
-
-export const setErrorPeersAction = () => ({ type: SET_ERROR_PEERS });
